+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to
 
 ### Changed
 
-- Used default J1 colors for `yarn j1-integration visualize-types` command
+- A single DataStore is now used for all dependency graphs executed in an
+  integration run.
+- Used default J1 colors for `yarn j1-integration visualize-types` command.
 
 ## 6.11.0 - 20201-07-14